Move RPC classes into a netty package
