+++ rskj-core/src/main/java/co/rsk/Start.java
@@ -40,9 +40,9 @@ import org.ethereum.core.*;
 import org.ethereum.core.genesis.BlockChainLoader;
 import org.ethereum.net.eth.EthVersion;
 import org.ethereum.net.server.ChannelManager;
-import org.ethereum.rpc.Web3HttpServer;
-import org.ethereum.rpc.JsonRpcWeb3FilterHandler;
-import org.ethereum.rpc.JsonRpcWeb3ServerHandler;
+import co.rsk.rpc.netty.Web3HttpServer;
+import co.rsk.rpc.netty.JsonRpcWeb3FilterHandler;
+import co.rsk.rpc.netty.JsonRpcWeb3ServerHandler;
 import org.ethereum.rpc.Web3;
 import org.ethereum.sync.SyncPool;
 import org.ethereum.util.BuildInfo;
