+++ rskj-core/src/main/java/co/rsk/rpc/netty/JsonRpcWeb3ServerHandler.java
@@ -16,7 +16,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.ethereum.rpc;
+package co.rsk.rpc.netty;
 
 import co.rsk.rpc.JsonRpcFilterServer;
 import co.rsk.rpc.ModuleDescription;
@@ -37,6 +37,7 @@ import io.netty.channel.ChannelHandler;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.SimpleChannelInboundHandler;
 import io.netty.handler.codec.http.*;
+import org.ethereum.rpc.Web3;
 import org.ethereum.rpc.exception.RskErrorResolver;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
